+++ rust-plugins/virtual/options.d.ts
@@ -1,0 +1,1 @@
+export type IPluginOptions = Record<string, any>;