+++ rust-plugins/react-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/plugin-react-components
 
+## 1.0.10
+
+### Patch Changes
+
+- feat(rust-plugin): optimization plugin
+
 ## 1.0.9
 
 ### Patch Changes